// Client/src/app/grant-call/grant-call-interfaces.ts
export class pointOfContact {
    fullName: string;
    email: string;
    mobile: string;
    designation: string;
    personId: number;
    grantContactId: number;
    isEmployee: boolean;
    grantCallId: number;
    rolodexId: number;
}